+++ pkg/commands/auth/root.go
@@ -1,0 +1,25 @@
+package auth
+
+import (
+	"io"
+
+	"github.com/fastly/cli/pkg/argparser"
+	"github.com/fastly/cli/pkg/global"
+)
+
+type RootCommand struct {
+	argparser.Base
+}
+
+const CommandName = "auth"
+
+func NewRootCommand(parent argparser.Registerer, g *global.Data) *RootCommand {
+	var c RootCommand
+	c.Globals = g
+	c.CmdClause = parent.Command(CommandName, "Manage stored Fastly API tokens and token policies")
+	return &c
+}
+
+func (c *RootCommand) Exec(_ io.Reader, _ io.Writer) error {
+	panic("unreachable")
+}