+++ pkg/commands/authtoken/authtoken_test.go
@@ -294,7 +294,10 @@ Expires at: 2021-06-15 23:00:00 +0000 UTC`
 
 func listTokenOutputVerbose() string {
 	return `Fastly API endpoint: https://api.fastly.com
-Fastly API token provided via config file (profile: user)
+Fastly API token provided via config file (auth: user)
+
+DEPRECATED: The 'auth-token' command tree will be removed in a future release. Use the Fastly API directly to manage API
+tokens.
 
 
 ID: 123