+++ pkg/commands/service/logging/cloudfiles/cloudfiles_integration_test.go
@@ -244,7 +244,7 @@ SERVICE  VERSION  NAME
 
 var listCloudfilesVerboseOutput = strings.TrimSpace(`
 Fastly API endpoint: https://api.fastly.com
-Fastly API token provided via config file (profile: user)
+Fastly API token provided via config file (auth: user)
 
 Service ID (via --service-id): 123