+++ pkg/commands/service/logging/datadog/datadog_integration_test.go
@@ -219,7 +219,7 @@ SERVICE  VERSION  NAME
 
 var listDatadogsVerboseOutput = strings.TrimSpace(`
 Fastly API endpoint: https://api.fastly.com
-Fastly API token provided via config file (profile: user)
+Fastly API token provided via config file (auth: user)
 
 Service ID (via --service-id): 123