+++ pkg/commands/service/logging/kafka/kafka_integration_test.go
@@ -266,7 +266,7 @@ SERVICE  VERSION  NAME
 
 var listKafkasVerboseOutput = strings.TrimSpace(`
 Fastly API endpoint: https://api.fastly.com
-Fastly API token provided via config file (profile: user)
+Fastly API token provided via config file (auth: user)
 
 Service ID (via --service-id): 123