+++ pkg/commands/service/logging/s3/s3_integration_test.go
@@ -298,7 +298,7 @@ SERVICE  VERSION  NAME
 
 var listS3sVerboseOutput = strings.TrimSpace(`
 Fastly API endpoint: https://api.fastly.com
-Fastly API token provided via config file (profile: user)
+Fastly API token provided via config file (auth: user)
 
 Service ID (via --service-id): 123