+++ pkg/commands/service/service_test.go
@@ -452,7 +452,7 @@ Baz   789  vcl   1               n/a
 
 var listServicesVerboseOutput = strings.TrimSpace(`
 Fastly API endpoint: https://api.fastly.com
-Fastly API token provided via config file (profile: user)
+Fastly API token provided via config file (auth: user)
 
 Service 1/3
 	ID: 123
@@ -590,7 +590,7 @@ Versions: 2
 
 var describeServiceVerboseOutput = strings.TrimSpace(`
 Fastly API endpoint: https://api.fastly.com
-Fastly API token provided via config file (profile: user)
+Fastly API token provided via config file (auth: user)
 
 Service ID (via --service-id): 123
 
@@ -681,7 +681,7 @@ Versions: 2
 
 var searchServiceVerboseOutput = strings.TrimSpace(`
 Fastly API endpoint: https://api.fastly.com
-Fastly API token provided via config file (profile: user)
+Fastly API token provided via config file (auth: user)
 
 ID: 123
 Name: Foo